+++ builtin/providers/aws/resource_aws_redshift_subnet_group.go
@@ -30,7 +30,8 @@ func resourceAwsRedshiftSubnetGroup() *schema.Resource {
 
 			"description": &schema.Schema{
 				Type:     schema.TypeString,
-				Required: true,
+				Optional: true,
+				Default:  "Managed by Terraform",
 			},
 
 			"subnet_ids": &schema.Schema{