+++ terraform/test-fixtures/input-interpolate-var/child/main.tf
@@ -1,0 +1,5 @@
+variable "list" { }
+resource "template_file" "temp" {
+  count     = "${length(split(",", var.list))}"
+  template  = "foo"
+}