+++ terraform/test-fixtures/validate-computed-module-var-ref/source/main.tf
@@ -1,0 +1,3 @@
+resource "aws_instance" "source" { }
+
+output "sourceout" { value = "${aws_instance.source.id}" }